requirements.txt
torch
numpy
tqdm
pyyaml
gymnasium
wandb
minigrid
tokenizers
rdkit
safetensors
selfies
tyro
hydra-core
pytest